+++ src/Agent/main.cpp
@@ -10,12 +10,7 @@ void SendTest(CTcpClient* client) {
 		printf("Message : ");
 		fgets(message, BUFFER_SIZE, stdin);
 
-		ST_PACKET_INFO stPacketSend;
-		stPacketSend.source = AGENT;
-		stPacketSend.destination = SERVER;
-		stPacketSend.type = REQUEST;
-		stPacketSend.opcode = OPCODE1;
-		stPacketSend.data = message;
+		ST_PACKET_INFO stPacketSend(AGENT, SERVER, REQUEST, OPCODE1, message);
 
 		std::tstring jsPacketSend;
 		core::WriteJsonToString(&stPacketSend, jsPacketSend);
